+++ cpp/src/Ice/Outgoing.cpp
@@ -381,7 +381,7 @@ IceInternal::Outgoing::finished(BasicStream& is)
     assert(_state <= StateInProgress);
     if(_remoteObserver)
     {
-        _remoteObserver->reply(is.b.size() - headerSize - 4);
+        _remoteObserver->reply(static_cast<Int>(is.b.size() - headerSize - 4));
     }
     _remoteObserver.detach();
